Log rotation added.
Even though logs are small (so far), did not want
to lose track of log rotation. Got that added.

Tested to confirm container rebuilt and started,
and service resumed running with original state and
successfully updated.

Did not test the log rotation, as we are a long
way from needing the defaults. But if it crashes
or a problem happens later, then it will show
up or trigger re-investigation.