Add configure endpoint
Added configure endpoint which allows for add, edit,
enable/disable, and delete of proxy endpoints.
These work dynamically. When you save an endpoint
the web server picks up the changes immediately,
no container down or action required on the
user's end.

This can be used to add endpoints to the web
server to point at other container web or api
endpoints on the localhost.

Tested with searxng to validate get and post work
through the proxy.