Enhance reset script.
Made changes to prevent unnecessary snapshot
rollover if the counters are already zero (i.e. if
a reset was already performed, or if no activity.)
This was causing some cruft in the snapshot
collection folder. Now the script will just log
a message that no reset was needed and exit.