Date cutoff fix on llm_collector snapshots
Since resets/snapshots of llm counts are done at or
right after midnight, timestamps and counts are
rolling the counts to the next calendar day, which
is incorrect.

Added a cutoff hour, and anything prior to that
will be rolled back to the prior calendar day.

This will enable the midnight or a bit later
snapshots to be accumulated to the prior day, while
still allowing mid day snapshots to roll up to the
current/correct day.